+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## NEXT 0.0.14(2022-11-15)
 
-- TODO add : doc of Command `>90%`
 - TODO: add `this.fn()` goto def, just def,don't support of ref ...
+- add: doc of Command to 218 ruler
 - add: progress of dev tools
 - add: support [Multi-line](https://www.autohotkey.com/docs/Scripts.htm#continuation-section) `%` flag
 - add: cache to find all ref of function